% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sircleAlluvial.R
\name{sircleAlluvial}
\alias{sircleAlluvial}
\title{sircleAlluvial}
\usage{
sircleAlluvial(filename, outputFilename = "SiRCle_Alluvial.pdf")
}
\arguments{
\item{filename}{Path to the input file}
}
\value{
alluvialPlt The plot generated
}
\description{
This function creates an alluvial plot from an input CSV.
}
