% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sircleDoro.R
\name{sircleDoro}
\alias{sircleDoro}
\title{sircleDoro}
\usage{
sircleDoro(
  regulons,
  matRNAseq,
  calcDiffBetweenCond = F,
  cond1 = NULL,
  cond2 = NULL,
  method = "rank",
  outputFileName = "SiRCle-dorethea_pred_tfactivities.csv"
)
}
\arguments{
\item{regulons}{Regulons from dorothea)}

\item{matRNAseq}{Matrix of RNAseq values}

\item{calcDiffBetweenCond}{Boolean to determine whether or not to calculate the difference in predicted activities}

\item{cond1}{A condition that we can filter the columns of the matrix by so the column must contain the value in Condition 1 (i.e. "KO")}

\item{cond2}{A condition that we can filter the columns of the matrix by so the column must contain the value in Condition 2 (i.e. "WT")}

\item{filename}{Path to the input file}
}
\value{
tfDf A data frame with the predicted activities
}
\description{
Uses dorothea and viper to predict the activity and activity change of transcription factors.
}
