% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sircleRCM.R
\name{sircleFET}
\alias{sircleFET}
\title{sircleFET}
\usage{
sircleFET(
  rcm,
  allGenes,
  genesOfInterest,
  outputFilename = "SiRCle-FET.tsv",
  groupLabels = c("MDE", "MDS", "TMDE", "TMDS", "TPDE", "TPDE_TMDS", "TPDS",
    "MDS_TMDE", "TPDS_TMDE", "MDE-ncRNA", "MDS-ncRNA")
)
}
\arguments{
\item{rcm}{Instance of RCM after running sircleRCM}

\item{allGenes}{Background of all genes}

\item{genesOfInterest}{Genes that you are interested in (i.e. from some gene list)}

\item{groupLabels}{\emph{Optional: } Labels from the groups  \strong{Default=c('MDE', 'MDS', 'TMDE', 'TMDS', 'TPDE', 'TPDE_TMDS', 'TPDS', 'MDS_TMDE', 'TPDS_TMDE', "MDE-ncRNA", "MDS-ncRNA")}}

\item{geneId}{GeneId column name in your dataframe}
}
\description{
Uses scircm to compute the regulatory clustering model.
}
