% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sircleTF.R
\name{sircleMotif}
\alias{sircleMotif}
\title{sircleMotif}
\usage{
sircleMotif(
  clusterGeneId,
  doroFile,
  sircleRCMFilename,
  TFPadjCol,
  TFValueCol,
  TargetPadjCol,
  TargetValueCol,
  RegulatoryLabel,
  TFInDataset = T,
  doroLevel = c("A"),
  clusters = c("MDS", "MDE", "TPDE", "TPDS"),
  outputFilename = "SiRCle-motif.csv",
  outputDir = ""
)
}
\arguments{
\item{clusterGeneId}{Column name with the gene identifier}

\item{doroFile}{Filename/path of the file from Dorothea which has A,B,C,D relationships}

\item{sircleRCMFilename}{Filename/path of the output from sircleRCM}

\item{TFPadjCol}{Column name of the padjusted value in the sircleRCMFilename file (i.e. your RNAseq, or your protein p column)}

\item{TFValueCol}{Column name of  rna value for the TF (could also use the proteomics logfc)}

\item{TargetPadjCol}{Column name of the padjusted value in the sircleRCMFilename file (i.e. your RNAseq, or your protein p column)}

\item{TargetValueCol}{Column name of  rna value for the TF (could also use the proteomics logfc)}

\item{RegulatoryLabel}{Column name of the regulatory labels}

\item{TFInDataset}{\emph{Optional: } Whether or not you require the TF to be expressed in the dataframe}

\item{doroLevel}{\emph{Optional: } level(s) of doro as a list \strong{Default=c("A")}}

\item{clusters}{\emph{Optional: } clusters to investigate \strong{Default=c("MDS", "MDE", "TPDE", "TPDS")}}

\item{outputFilename}{\emph{Optional: } Filename for the output file.}

\item{outputDir}{\emph{Optional: } Full path for output directory.}
}
\value{
scimo an instance of the scimo package
}
\description{
Uses dorothea to find TFs enriched in clusters
}
