% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sircleTF.R
\name{sircleMotifAddInfo}
\alias{sircleMotifAddInfo}
\title{sircleMotifAddInfo}
\usage{
sircleMotifAddInfo(
  scimo,
  motifDf,
  otherFilename,
  TFNameColumn,
  TFValueColumn,
  valueLabel,
  outputFilename = "SiRCle-motif-updated.csv"
)
}
\arguments{
\item{scimo}{The return value from sircleMotif (this is an instance of the scimo package.)}

\item{motifDf}{Dataframe output by the previous analysis}

\item{otherFilename}{Filename of the file from which the new data come from}

\item{TFNameColumn}{Column name of the TF within the otherfile (we use the TFs in this column to match with those in motif DF so they need to be in the same format!)}

\item{valueLabel}{Label to append to the TFValue column name so that you know where the source came from.}

\item{outputFilename}{Filename of the new file to be saved}

\item{TFVlaueColumn}{Column name with the TF values within the other file}
}
\value{
dm_df A data frame with the predicted activities and the added column
}
\description{
Adds info to the sircleMotif dataframe (i.e. TF activtieis or protein values.)
}
