% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sircleORA.R
\name{sircleORAHuman}
\alias{sircleORAHuman}
\title{sircleORAHuman}
\usage{
sircleORAHuman(
  filename,
  entrezId,
  regLabels = "RegulatoryLabels",
  emptyRegLabel = "",
  fileType = "pdf",
  minGSSize = 10,
  qvalueCutoff = 0.2
)
}
\arguments{
\item{filename}{Path to the input file}

\item{entrezId}{Column name for the entrez ID}

\item{regLabels}{\emph{Optional: } regLabels The label of the column with the regulatory labels \strong{default: "RegulatoryLabels"}}

\item{emptyRegLabel}{\emph{Optional: } emptyRegLabel The label of the empty regulatory group \strong{default: ""}}

\item{fileType}{\emph{Optional: } fileType Output file type for the figures one of: "svg", "pdf", "png" \strong{default: "pdf"}}

\item{minGSSize}{\emph{Optional: } minimum group size in ORA \strong{default: 10}}

\item{qvalueCutoff}{\emph{Optional: } q value cutoff from ORA \strong{default: 0.2}}
}
\description{
Uses encrichGo to run ORA on each of the clusters as defined by the regulatory labels (ON HUMAN)
}
