% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sircleORA.R
\name{sircleORAMouse}
\alias{sircleORAMouse}
\title{sircleORAMouse}
\usage{
sircleORAMouse(filename, regLabels = "RegulatoryLabels", fileType = "pdf")
}
\arguments{
\item{filename}{Path to the input file}

\item{rcm}{Instance of RCM after running sircleRCM}

\item{envPath}{\emph{Optional: } fileType Output file type for the figures one of: "svg", "pdf", "png" \strong{default: "pdf"}}
}
\description{
Uses encrichGo to run ORA on each of the clusters as defined by the regulatory labels (on MOUSE)
}
