% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sircleTFVis.R
\name{sircleTFVis}
\alias{sircleTFVis}
\title{sircleTFVis}
\usage{
sircleTFVis(
  filename,
  colorCol = "q-value",
  fillNAVal = NA,
  plotBar = T,
  plotEman = T,
  titleStr = ""
)
}
\arguments{
\item{filename}{Column name with the gene identifier}

\item{colorCol}{\emph{Optional:} Column name to use for colouring \strong{Default: "q-value"}}

\item{fillNAVal}{\emph{Optional:} Value to fill None values with \strong{Default: 0.0}}

\item{plotBar}{\emph{Optional:} Whether or not to plot the bar charts \strong{Default: T}}

\item{plotEman}{\emph{Optional:} Whether or not to plot the connection plot \strong{Default: T}}

\item{titleStr}{\emph{Optional:} String for the title \strong{Default: ""}}
}
\value{
scimo an instance of the scimo package
}
\description{
Creates visualisations for the TF analysis.
}
