% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sircleRCM.R
\name{sircleVAE}
\alias{sircleVAE}
\title{sircleVAE}
\usage{
sircleVAE(
  rcm,
  df,
  geneId,
  configVAEFile,
  plotVAE = T,
  plottingCols = NULL,
  figType = "pdf"
)
}
\arguments{
\item{rcm}{Instance of RCM after running sircleRCM}

\item{df}{Dataframe with the values we want as our input to the VAE (these should all be numeric except for the gene ID column!)}

\item{geneId}{Gene ID column name in you dataframe}

\item{configVAEFile}{Filename of the config.json (see scivae for description.)}

\item{plotVAE}{\emph{Optional: } Whether or not you want to plot (and save) the figures for the VAE \strong{Default=T}}

\item{plottingCols}{\emph{Optional: } Columns for plotting in your dataset - if you don't select these, and you choose to plot all your columns will be used! \strong{Default=NULL}}

\item{figType}{\emph{Optional: } File ending for the saved figures \strong{Default="pdf"}}
}
\value{
rcm an instance of the rcm package
}
\description{
Uses scircm to compute the regulatory clustering model.
}
